More some helpers to packages/core for the cli and some more types
