+++ packages/core/src/browser.ts
@@ -1,0 +1,6 @@
+/**
+ * Browser-safe exports for the core package. These can safely be used
+ * in browser environments like `webview-ui`.
+ */
+
+export * from "./message-utils/index.js"
