+++ packages/core/src/cli.ts
@@ -1,3 +1,6 @@
-export * from "./custom-tools/index.js"
+/**
+ * Cli-safe exports for the core package.
+ */
+
 export * from "./debug-log/index.js"
 export * from "./message-utils/index.js"
