+++ packages/types/src/embedding.ts
@@ -1,0 +1,22 @@
+export type EmbedderProvider =
+	| "openai"
+	| "ollama"
+	| "openai-compatible"
+	| "gemini"
+	| "mistral"
+	| "vercel-ai-gateway"
+	| "bedrock"
+	| "openrouter" // Add other providers as needed.
+
+export interface EmbeddingModelProfile {
+	dimension: number
+	scoreThreshold?: number // Model-specific minimum score threshold for semantic search.
+	queryPrefix?: string // Optional prefix required by the model for queries.
+	// Add other model-specific properties if needed, e.g., context window size.
+}
+
+export type EmbeddingModelProfiles = {
+	[provider in EmbedderProvider]?: {
+		[modelId: string]: EmbeddingModelProfile
+	}
+}
