+++ packages/types/src/index.ts
@@ -4,6 +4,7 @@ export * from "./codebase-index.js"
 export * from "./context-management.js"
 export * from "./cookie-consent.js"
 export * from "./custom-tool.js"
+export * from "./embedding.js"
 export * from "./events.js"
 export * from "./experiment.js"
 export * from "./followup.js"
