+++ src/core/tools/BaseTool.ts
@@ -1,3 +1,5 @@
+import type { ToolName, ToolProtocol } from "@roo-code/types"
+
 import { Task } from "../task/Task"
 import type {
 	ToolUse,
@@ -7,7 +9,6 @@ import type {
 	AskApproval,
 	NativeToolArgs,
 } from "../../shared/tools"
-import type { ToolName, ToolProtocol } from "@roo-code/types"
 
 /**
  * Callbacks passed to tool execution
