+++ src/core/tools/accessMcpResourceTool.ts
@@ -1,7 +1,9 @@
-import { ClineAskUseMcpServer } from "../../shared/ExtensionMessage"
+import type { ClineAskUseMcpServer } from "@roo-code/types"
+
 import type { ToolUse } from "../../shared/tools"
 import { Task } from "../task/Task"
 import { formatResponse } from "../prompts/responses"
+
 import { BaseTool, ToolCallbacks } from "./BaseTool"
 
 interface AccessMcpResourceParams {
