+++ src/shared/core.ts
@@ -1,0 +1,1 @@
+export * from "@roo-code/core/browser"
