+++ src/shared/todo.ts
@@ -1,4 +1,5 @@
 import { ClineMessage } from "@roo-code/types"
+
 export function getLatestTodo(clineMessages: ClineMessage[]) {
 	const todos = clineMessages
 		.filter(
@@ -15,6 +16,7 @@ export function getLatestTodo(clineMessages: ClineMessage[]) {
 		.filter((item) => item && item.tool === "updateTodoList" && Array.isArray(item.todos))
 		.map((item) => item.todos)
 		.pop()
+
 	if (todos) {
 		return todos
 	} else {
